pyproject.toml
[tool.poetry]
name = "azure-create-policy-exceptions"
version = "0.1.0"
description = ""
authors = ["githubofkrishnadhas <krishnadhasnk1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-mgmt-policyinsights = "^1.0.0"
azure-identity = "^1.17.1"
python-dotenv = "^1.0.1"
azure-mgmt-resource = "^23.1.1"
azure-mgmt-resourcegraph = "^8.0.0"
pydantic = "^2.8.2"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
azure-core = "^1.30.2"
streamlit = "1.36.0"
streamlit-msal = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
